\name{safi-package}
\alias{safi-package}
\alias{safi}
\docType{package}
\title{
safi
}
\description{
Sensitivity analysis for functional inputs.
}

\author{
Jana Fruth, Malte Jastrow
}

\keyword{ package }

\references{Fruth, J.; Roustant, O.; Kuhnt, S. (2014) Sequential designs for sensitivity analysis of functional inputs in computer experiments, \emph{Reliability Engineering & System Safety}, doi: 10.1016/j.ress.2014.07.018, preprint on HAL: \url{http://hal.archives-ouvertes.fr/hal-00943509}.
}